Fix benchmark suite
